ai-cv-assistant/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
python-dotenv
httpx
openai
